Oops, missed a couple.
Reviewed-on: https://code-review.googlesource.com/c/re2/+/43032

+++ re2/testing/dfa_test.cc
@@ -295,7 +295,7 @@ ReverseTest reverse_tests[] = {
 
 TEST(DFA, ReverseMatch) {
   int nfail = 0;
-  for (int i = 0; i < arraysize(reverse_tests); i++) {
+  for (size_t i = 0; i < arraysize(reverse_tests); i++) {
     const ReverseTest& t = reverse_tests[i];
     Regexp* re = Regexp::Parse(t.regexp, Regexp::LikePerl, NULL);
     ASSERT_TRUE(re != NULL);
@@ -350,7 +350,7 @@ CallbackTest callback_tests[] = {
 
 TEST(DFA, Callback) {
   int nfail = 0;
-  for (int i = 0; i < arraysize(callback_tests); i++) {
+  for (size_t i = 0; i < arraysize(callback_tests); i++) {
     const CallbackTest& t = callback_tests[i];
     Regexp* re = Regexp::Parse(t.regexp, Regexp::LikePerl, NULL);
     ASSERT_TRUE(re != NULL);
